Small layout changes, and added a reset button.
Added a navbar to the site, as I plan on adding more things other than
just a sudoku (otherwise, it'd be html5sudoku, not experiments).

Tiny layout changes with the buttons etc, and added a new reset button,
that will reset the board back to what the starting board was.