Added partials and path changes
Added in partials so we have the same html structure for all pages, and
different controllers for each page.

Also, changed the node server so that it automatically goes to
index.html if pointed to a directory, and if pointed to a directory that
doesn't exist, it will go down until it either finds something or hits
the end. If it tries to find a file, it'll still 404.