Rearranging reliances and re-architecting require.
Rearranged where the dependencies are. They are now managed by bower
(just run bower install to update), and are stored in app/lib.

Also, rearchitected it all to use require (because everyone loves
require), and made sure only things in each file are the things we need
in that file.