Tests and persistency
Well, kinda. I added 2 utility functions to allow saving and loading,
but what they do is turn the board into a string, and a string into the
board. So, they don't *actually* persist, but it allows me to do various
other things really easily. For example, saving the board in the url,
cookie or html5 storage. I'll get to that.

Also, added tests for these utility functions, and updated the
controller tests (kinda).