Few visual fixes and enhancements
Fixed the annoying FF scrollbar bug (showing a scrollbar when hovering
over the sudoku board).

Fixed the other annoying FF/safari bug where the background of the cell
overlapped the border of the tr/col.

Tried to fix the col:hover not working in FF (didn't work).

And made the table slightly rounded. Looks a bit nicer that way :-)