Appearance changes
Now there are 2 lots of buttons (marked control and cheat) to
distinguish between different types of controls. Also, it means it takes
a bit more space on wide-screen now, which is good.

Also, fixed the stupid overlay. Now it has a container div that is
shrink-wrapped, and both the table and overlay are placed in the right
point. It seems to looks good on all screens sizes that I can see.