Moved the status info
The info as to whether it's ok so far or not now shows as a table
overlay in xs mode. Not in anything else, as there is no real need, I
just thought it'd be nicer that way, then we can hide the buttons away
when not being used (not done yet) and just makes a much nicer mobile
interface.