Changed pencil marks and checking
Pencil marks are stored as an array of values, rather than an array of
binary values.

Also, the interface now shows you what is wrong based on the correct
board. It's a quicker way to check that it's ok, and as there is only 1
board, then at least we can know that if it's not in the correct board,
it's not correct. Plus, user feedback said that's what they expected it
to do.