Added rater and single cell solving
Added in solving for cells that only have 1 option. This is actually
enough, interestingly, to solve easy sudoku. Once it solves that cell,
it removes the number from all cells that have that number in their
pencil marks (in their column, row and box obviously).

Also, added a rater which uses the human solver to give a sudoku a
rating.

Currently, the rating will be a max of 81 (although that's impossible to
get with the current solver). Anything less than 81 minus the original
number of filled in cells means it can't be solved. Eventually, the
rating will just be 0 to undefined, with 0 being the easiest, going
upwards with the only limit being undefined which means it's unsolvable
with the current rules.