Fixed 2 small bugs
The correct board wasn't being empties on clear, which meant that when
dealing with an empty board, it was comparing to the last correct board,
meaning it was, most of the time, incorrect.

Also, fixed a bug with the pencil marks. I hadn't changed the clear
functions way of clearing pencil marks. Caused issues.