Modified how the html gets data
It now uses the actual Board data structure rather than the underlying
array. This is much more sensible (obviously) and allows me to change
the underlying structure. Because I think it might be incredibly
inefficient just now.

Also, fixed a bug with the pencil marks being still added to the old way
of defining them.