Added multiple web workers and background generation
There are currently 5 web workers that generate sudoku now, and pass
them back to the controller, which stores them and waits for the user to
request a new game. When they do, it checks if we have one already
generated, and uses that one. Otherwise, it does the old style
generation.

Next stop...ranking!