Added client side ratings.
So, when we get the board back, we check the rating and stick the board
in the correct array. Then when we want a new game, we check the current
difficulty setting and give an appropriate one, or generate a new one.

Not sure what to do about generating an appropriate one as currently it
just generates a random one which probably doesn't fit the correct
difficulty setting.